requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
black==24.10.0
click==8.1.8
colorama==0.4.6
decouple==0.0.7
Django==5.1.5
django-admin-interface==0.29.4
django-colorfield==0.11.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
PyJWT==2.9.0
python-dotenv==1.0.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2024.2
